Update installation.md
You will need to create the data base First. See your web hosting service for instructions on how. Use the name of the database you created, the hostname, the user name and the password. The install will need these to access it.

(if possible please provide a link to a more complete set of instructions. The Installation and Update forum only  helps some what.)